pyproject.toml
[project]
name = "photo-management"
version = "2024.1"
description = "Tools for managing a simple photo library"
readme = "README.md"
authors = [
    { name = "Sam H", email = "samhutchins@users.noreply.github.com" }
]
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "pyexiftool>=0.5.6",
]

[project.urls]
Repository = "https://github.com/samhutchins/photo_management"

[project.scripts]
manage-photos = "photo_management.photo_management:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.8.3",
    "types-python-dateutil>=2.9.0.20241206",
]
